refactor: divide http & websocket connection to multi modules
