+++ src/graia/ariadne/exception.py
@@ -4,6 +4,7 @@
 class AriadneConfigureError(Exception):
     """配置 Ariadne 时给出的参数包含有错误."""
 
+
 class InvalidEventTypeDefinition(Exception):
     """不合法的事件类型定义."""
